Split out AstFold into separate module
Improve unnesting
